audit_service/requirements.txt
# fastapi==0.111.0
# uvicorn[standard]==0.23.2
# sqlalchemy==2.0.22
# databases==0.7.2
# asyncpg==0.27.1
# pydantic==2.2.2
# python-dotenv==1.0.0

fastapi==0.111.0
uvicorn[standard]==0.23.2
# sqlalchemy==1.4.46
# databases==0.7.0
# sqlalchemy==2.0.22
# databases==0.8.0
sqlalchemy>=1.4.42,<1.5
databases==0.8.0
# asyncpg==0.27.0
asyncpg>=0.30.0
pydantic==2.12.4
python-dotenv==1.0.0

